Cleans up tag processing and coro usage

+++ IDHANMigration/src/93-tag_counts.sql
@@ -1,6 +1,6 @@
 CREATE TABLE tag_counts
 (
-    tag_id        BIGINT REFERENCES tags (tag_id),
+    tag_id        INTEGER REFERENCES tags (tag_id),
     storage_count INTEGER NOT NULL DEFAULT 0,
     display_count INTEGER NOT NULL DEFAULT 0,
     tag_domain_id SMALLINT REFERENCES tag_domains (tag_domain_id),
@@ -9,13 +9,13 @@ CREATE TABLE tag_counts
 
 CREATE TABLE total_tag_counts
 (
-    tag_id        BIGINT REFERENCES tags (tag_id),
+    tag_id        INTEGER REFERENCES tags (tag_id),
     storage_count INTEGER NOT NULL DEFAULT 0,
     display_count INTEGER NOT NULL DEFAULT 0,
     PRIMARY KEY (tag_id)
 );
 
-CREATE OR REPLACE FUNCTION update_tag_counts(target_tag_id BIGINT) RETURNS VOID AS
+CREATE OR REPLACE FUNCTION update_tag_counts(target_tag_id INTEGER) RETURNS VOID AS
 $$
 BEGIN
     -- Lock the tag_counts table to prevent concurrent updates
@@ -24,8 +24,9 @@ BEGIN
     -- Update/Insert counts using a single UPSERT operation
     INSERT INTO total_tag_counts (tag_id, storage_count, display_count)
     SELECT target_tag_id,
-           COUNT(DISTINCT record_id) FILTER (WHERE atm.tag_id = target_tag_id)                             AS storage_count,
-           COUNT(DISTINCT record_id) FILTER (WHERE COALESCE(atm.ideal_tag_id, atm.tag_id) = target_tag_id) AS display_count
+           COUNT(DISTINCT record_id) FILTER (WHERE atm.tag_id = target_tag_id)   AS storage_count,
+           COUNT(DISTINCT record_id)
+           FILTER (WHERE COALESCE(atm.ideal_tag_id, atm.tag_id) = target_tag_id) AS display_count
     FROM active_tag_mappings atm
     WHERE atm.tag_id = target_tag_id
        OR COALESCE(atm.ideal_tag_id, atm.tag_id) = target_tag_id
@@ -35,10 +36,11 @@ BEGIN
 
     -- Update/Insert counts per domain using a single UPSERT operation
     INSERT INTO tag_counts (tag_id, tag_domain_id, storage_count, display_count)
-    SELECT target_tag_id                                                                                   AS tag_id,
-           atm.tag_domain_id                                                                               AS tag_domain_id,
-           COUNT(DISTINCT record_id) FILTER (WHERE atm.tag_id = target_tag_id)                             AS storage_count,
-           COUNT(DISTINCT record_id) FILTER (WHERE COALESCE(atm.ideal_tag_id, atm.tag_id) = target_tag_id) AS display_count
+    SELECT target_tag_id                                                         AS tag_id,
+           atm.tag_domain_id                                                     AS tag_domain_id,
+           COUNT(DISTINCT record_id) FILTER (WHERE atm.tag_id = target_tag_id)   AS storage_count,
+           COUNT(DISTINCT record_id)
+           FILTER (WHERE COALESCE(atm.ideal_tag_id, atm.tag_id) = target_tag_id) AS display_count
     FROM active_tag_mappings atm
     WHERE (atm.tag_id = target_tag_id OR COALESCE(atm.ideal_tag_id, atm.tag_id) = target_tag_id)
     GROUP BY atm.tag_domain_id
@@ -48,7 +50,7 @@ BEGIN
 END;
 $$ LANGUAGE plpgsql;
 
-CREATE FUNCTION add_count(tag_id_i BIGINT, ideal_tag_id_i BIGINT, tag_domain_id_i SMALLINT) RETURNS VOID AS
+CREATE FUNCTION add_count(tag_id_i INTEGER, ideal_tag_id_i INTEGER, tag_domain_id_i SMALLINT) RETURNS VOID AS
 $$
 BEGIN
 
@@ -68,14 +70,17 @@ BEGIN
 END;
 $$ LANGUAGE plpgsql;
 
-CREATE FUNCTION remove_count(tag_id_i BIGINT, ideal_tag_id_i BIGINT, tag_domain_id_i SMALLINT) RETURNS VOID AS
+CREATE FUNCTION remove_count(tag_id_i INTEGER, ideal_tag_id_i INTEGER, tag_domain_id_i SMALLINT) RETURNS VOID AS
 $$
 BEGIN
 
     LOCK TABLE tag_counts IN EXCLUSIVE MODE;
 
     UPDATE tag_counts SET storage_count = storage_count - 1 WHERE tag_id = tag_id_i AND tag_domain_id = tag_domain_id_i;
-    UPDATE tag_counts SET display_count = display_count - 1 WHERE tag_id = COALESCE(ideal_tag_id_i, tag_id_i) AND tag_domain_id = tag_domain_id_i;
+    UPDATE tag_counts
+    SET display_count = display_count - 1
+    WHERE tag_id = COALESCE(ideal_tag_id_i, tag_id_i)
+      AND tag_domain_id = tag_domain_id_i;
 END;
 $$ LANGUAGE plpgsql;
